Refactor application into modular packages
