+++ internal/output/mode.go
@@ -1,0 +1,11 @@
+package output
+
+// Mode defines how results are presented.
+type Mode int
+
+const (
+	// ModeCLI prints endpoints to stdout.
+	ModeCLI Mode = iota
+	// ModeHTML renders endpoints into an HTML report.
+	ModeHTML
+)
